requirements.txt
# MalVec Dependencies
# Python 3.11+ required

# ====================
# EMBER Dataset (Phase 2)
# NOTE: EMBER is not on PyPI, install separately:
#   pip install git+https://github.com/elastic/ember.git
# ====================
# ember installed from git (see above)

# ====================
# Embeddings (Phase 3)
# ====================
sentence-transformers>=2.2.0
torch>=2.0.0

# ====================
# Vector Database (Phase 4)
# ====================
faiss-cpu>=1.7.4

# ====================
# Data & Utils
# ====================
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# ====================
# Static Analysis (Phase 7 - Binary Mode)
# ====================
pefile>=2023.2.7
lief>=0.13.0

# ====================
# Testing
# ====================
pytest>=7.4.0
pytest-cov>=4.1.0

# ====================
# Visualization
# ====================
matplotlib>=3.7.0
plotly>=5.14.0

# ====================
# Polish & UX (Phase 9)
# ====================
rich>=10.0.0           # Beautiful terminal output
pyyaml>=5.4.0          # YAML configuration support
